feat(shop) : 상점 기능 추가
- 상점 구매 추가
- 상점 구매 시 수량 지정 가능하게 기능 추가
- 상점 품목 추가
- 방지권 사용을 위한 체크박스 추가
- 돈이 없을 경우 등 명령어 수행이 불가할 경우 나타날 오류 메세지 추가

+++ BaggerClicker/Assets/Scripts/SO/Sword.cs
@@ -30,8 +30,8 @@ public class Sword : ScriptableObject, IIndexable
     [SerializeField] private Sprite swordSprite;
     [SerializeField] private int enhanceCount;
     [SerializeField] private int enhanceProbability;
-    [SerializeField] private int enhanceCost;
-    [SerializeField] private int sellValue;
+    [SerializeField] private ulong enhanceCost;
+    [SerializeField] private ulong sellValue;
     [SerializeField] private Sword nextSword;
     [SerializeField] private Sword first;
